--- android/run-local-test/local.json
@@ -1,22 +1,0 @@
-{
-    "user": "BROWSERSTACK_USERNAME",
-    "key": "BROWSERSTACK_ACCESS_KEY",
-    "capabilities": {
-        "app" : "bs://<app-id>",
-        "bstack:options": {
-            "projectName": "Pytest Android Project",
-            "buildName": "browserstack-build-1",
-            "sessionName": "BStack local_test",
-            "local": "true",
-            "debug": "true",
-            "networkLogs": "true"
-        }
-    },
-    "environments": [
-        {
-            "platformName": "android",
-            "platformVersion": "9.0",
-            "deviceName": "Google Pixel 3"
-        }
-    ]
-}
