--- ios/run-parallel-test/parallel.json
@@ -1,27 +1,0 @@
-{
-    "user": "BROWSERSTACK_USERNAME",
-    "key": "BROWSERSTACK_ACCESS_KEY",
-    "capabilities": {
-        "app" : "bs://<app-id>",
-        "bstack:options": {
-            "projectName": "Pytest iOS Project",
-            "buildName": "browserstack-build-1",
-            "sessionName": "BStack parallel_test",
-            "local": false,
-            "debug": "true",
-            "networkLogs": "true"
-        }
-    },
-    "environments": [
-        {
-            "platformName": "ios",
-            "platformVersion": "16",
-            "deviceName": "iPhone 14"
-        },
-        {
-            "platformName": "ios",
-            "platformVersion": "15",
-            "deviceName": "iPhone 11"
-        }
-    ]
-}
